C TRAVEL
This travel website project offers a seamless and visually appealing user experience for booking travel destinations. Built with HTML, CSS, and JavaScript, it utilizes the Bootstrap framework for responsive design and Font Awesome for icons. The site includes sections for booking travel, showcasing packages, displaying services, and a gallery. Users can navigate easily through the site to explore travel options, view packages with pricing and ratings, and utilize various services offered by the travel agency.

IMPORTANT!!!

I didn't add the JS files and PHP files of this project because GitHub pages doesn't support php. And also this need to be connect to a database server. Because of that this website doesn't work properly.